requirements.txt
moviepy
opencv-python
pillow
pygame